Introduce MP.Utility that contains simple utility methods
